# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/users/register": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "User registration",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "properties": {
                                            "login": {
                                                "type": "string",
                                                "example": "SomeLogin"
                                            },
                                            "password": {
                                                "type": "string",
                                                "example": "password123"
                                            },
                                            "first_name": {
                                                "type": "string",
                                                "example": "Michael"
                                            },
                                            "last_name": {
                                                "type": "string",
                                                "example": "Philips"
                                            },
                                            "birthdate": {
                                                "type": "string",
                                                "format": "date",
                                                "example": "14.02.2004"
                                            }
                                        },
                                        "type": "object"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Registered",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "null",
                                            "example": "null"
                                        },
                                        "result": {
                                            "properties": {
                                                "message": {
                                                    "type": "string",
                                                    "example": "Вы успешно зарегестрировались!"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "This user already exists!",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "properties": {
                                                "message": {
                                                    "type": "string",
                                                    "example": "Пользователь с таким логином уже существует"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "result": {
                                            "type": "null",
                                            "example": "null"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/users/login": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Login",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "properties": {
                                            "login": {
                                                "type": "string",
                                                "example": "SomeLogin"
                                            },
                                            "password": {
                                                "type": "string",
                                                "example": "password123"
                                            }
                                        },
                                        "type": "object"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Logged In",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "null",
                                            "example": "null"
                                        },
                                        "result": {
                                            "properties": {
                                                "message": {
                                                    "type": "string",
                                                    "example": "Вы успешно вошли в аккаунт!"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "User is not found!",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "properties": {
                                                "message": {
                                                    "type": "string",
                                                    "example": "Логин или пароль были введены неверно!"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "result": {
                                            "type": "null",
                                            "example": "null"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    }
}